Integration Domains: add unit tests
This commit implements unit tests for the Integration Domains app
to make sure the adapter (serializer), model, and view classes work
as expected. The unit tests are using factory-boy library for
creating the test data.

Coverage: test domain __str__() method and adapter class so that
the values for each of the fields in the serializer match the
values for each of the fields on the model. Views unit tests are
also added to make sure the domains API endpoints (GET, POST,
and DELETE) work as expected.

Signed-off-by: Francisco Trivino <ftrivino@redhat.com>