chore(GitHubActions): Extract Dockerfile to repository and build it
Move the Dockerfile meant as test container to repo
and build it using the path to the filename.
push built container to the github registry
when it is a PR and to the quay.io with latest
tag when it is run on main branch.
Do all this only when it is not fork.

Resolves: https://github.com/freeipa/ipa-tuura/issues/38

Signed-off-by: Tibor Dudlák <tdudlak@redhat.com>