[wsgi] Add project path to the sys.path
When the WSGI server loads, Django needs to import the settings module by using
the DJANGO_SETTINGS_MODULE environment variable to locate the appropriate settings
module. It must contain the dotted path to the settings module.

This commit is adding the path to the settings module.

Related: https://github.com/freeipa/ipa-tuura/issues/29

Signed-off-by: Francisco Trivino <ftrivino@redhat.com>