Integration domain: users_dn is a full DN
The config parameter users_dn is documented as
"Optional full DN of LDAP tree where users are"
but the current implementation is not consistent with the description.

Make this parameter mandatory, and expect it to contain the full
DN of the users tree, for instance:
ou=users,dc=example,dc=com

With this new behavior, no need to build the user dn from
relative dn + suffix.