Domain: act as a singleton class
Override the primary key of the domain to 1 in order to force the DB to
save just a single domain. This will raise an error when registering a
domain when there's one already registered.

In order to check if the domain addition is valid, we check if there's
any domain already registered and raise an error response accordingly.

Signed-off-by: Antonio Torres <antorres@redhat.com>