Add endpoint for GSSAPI Authentication
Add the `login_kerberos` endpoint for handling GSSAPI authentications.
mod_auth_gssapi and gssproxy are included as dependencies. Additional
steps are included to the IPA domain addition, such as the addition of
the HTTP service and keytab retrieval. Additionally, `login_password`
endpoint is provided as well, which requests a ticket using the user and
password passed with the client request.

Signed-off-by: Antonio Torres <antorres@redhat.com>